Add ability to launch server with all features *without* launching Cypress
In order for hooks and transactions to work, the LaunchesCypress class
has to be used to launch the Rails server. This adds the ability to
bypass the launching of Cypress (e.g. because it needs to run outside
the container that Rails runs in).

It also adds the ability to specify the hostname for the Rails server to
listen on via CYPRESS_RAILS_HOST.