support Rails 6.1.4 by making shared_connection_pool optional
this commit adds a CYPRESS_RAILS_SHARED_CONNECTION_POOL env var which defaults
to true. if you set it to false, you can run cypress-rails with Rails 6.1.4 --
but not, of course, with a shared connection pool as well.

this is obviously not a perfect solution. my client's app doesn't use a shared
connection pool, so it solves the problem for my limited use case, but it may
only be useful for this project as a starting point, or as a stopgap, for other
users who don't have shared connection pools to consider either.